Modularize components into separate files
