+++ src/App.js
@@ -5,17 +5,7 @@ import type { Track, EncodedTrack } from "./types";
 import Tone from "tone";
 
 import React, { Component } from "react";
-import {
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  FABButton,
-  Icon,
-  Slider,
-  Switch,
-} from "react-mdl";
+import { Button } from "react-mdl";
 
 import "./App.css";
 import "react-mdl/extra/css/material.light_blue-pink.min.css";
@@ -23,158 +13,10 @@ import "react-mdl/extra/material.js";
 
 import * as sequencer from "./sequencer";
 import * as model from "./model";
-import samples from "./samples.json";
 
-
-class SampleSelector extends Component {
-  state: {
-    open: boolean,
-  };
-
-  constructor(props) {
-    super(props);
-    this.state = {open: false};
-  }
-
-  open = (event) => {
-    event.preventDefault();
-    this.setState({open: true});
-  };
-
-  close = () => {
-    this.setState({open: false});
-  };
-
-  onChange = (event) => {
-    const {id, onChange} = this.props;
-    onChange(id, event.target.value);
-    this.close();
-  };
-
-  render() {
-    const {current} = this.props;
-    const {open} = this.state;
-    if (open) {
-      return (
-        <select autoFocus value={current} onChange={this.onChange} onBlur={this.close}>{
-          samples.map((sample, i) => {
-            return <option key={i}>{sample}</option>;
-          })
-        }</select>
-      );
-    } else {
-      return <a href="" onClick={this.open}>{current}</a>;
-    }
-  }
-}
-
-function TrackListView({
-  tracks,
-  currentBeat,
-  toggleTrackBeat,
-  setTrackVolume,
-  updateTrackSample,
-  muteTrack,
-  clearTrack,
-  deleteTrack,
-}) {
-  return (
-    <tbody>{
-      tracks.map((track, i) => {
-        return (
-          <tr key={i}className="track">
-            <th>
-              <SampleSelector id={track.id} current={track.name} onChange={updateTrackSample} />
-            </th>
-            <td className="vol">
-              <Slider min={0} max={1} step={.1} value={track.vol}
-                onChange={event => setTrackVolume(track.id, parseFloat(event.target.value))} />
-            </td>
-            <td className="mute">
-              <Switch defaultChecked={!track.muted} onChange={event => muteTrack(track.id)} />
-            </td>
-            {
-              track.beats.map((v, beat) => {
-                const beatClass = v ? "active" : beat === currentBeat ? "current" : "";
-                return (
-                  <td key={beat} className={`beat ${beatClass}`}>
-                    <a href="" onClick={(event) => {
-                      event.preventDefault();
-                      toggleTrackBeat(track.id, beat);
-                    }} />
-                  </td>
-                );
-              })
-            }
-            <td>
-              {track.beats.some(v => v) ?
-                <a href="" title="Clear track" onClick={event => {
-                  event.preventDefault();
-                  clearTrack(track.id);
-                }}><Icon name="delete"/></a> :
-                <Icon className="disabled-icon" name="delete"/>}
-              <a href="" title="Delete track" onClick={event => {
-                event.preventDefault();
-                deleteTrack(track.id);
-              }}><Icon name="delete_forever"/></a>
-            </td>
-          </tr>
-        );
-      })
-    }</tbody>
-  );
-}
-
-function Controls({bpm, updateBPM, playing, start, stop, addTrack, share}) {
-  const onChange = event => updateBPM(parseInt(event.target.value, 10));
-  return (
-    <tfoot className="controls">
-      <tr>
-        <td style={{textAlign: "right"}}>
-          <FABButton mini colored onClick={addTrack} title="Add new track">
-            <Icon name="add" />
-          </FABButton>
-        </td>
-        <td />
-        <td>
-          <FABButton mini colored onClick={playing ? stop : start}>
-            <Icon name={playing ? "stop" : "play_arrow"} />
-          </FABButton>
-        </td>
-        <td colSpan="2" className="bpm">
-          BPM <input type="number" value={bpm} onChange={onChange} />
-        </td>
-        <td colSpan="13">
-          <Slider min={30} max={240} value={bpm} onChange={onChange} />
-        </td>
-        <td colSpan="2">
-          <FABButton mini onClick={share} title="Share">
-            <Icon name="share" />
-          </FABButton>
-        </td>
-      </tr>
-    </tfoot>
-  );
-}
-
-function ShareDialog({hash, closeDialog}) {
-  return (
-    <Dialog open>
-      <DialogTitle>Share</DialogTitle>
-      <DialogContent>
-        <p>Send this link to your friends so they can enjoy your piece:</p>
-        <p className="share-link" style={{textAlign: "center"}}>
-          <a className="mdl-button mdl-js-button mdl-button--colored"
-            href={"#" + hash} onClick={event => event.preventDefault()}>Link</a>
-        </p>
-        <p>Right-click, <em>Copy link address</em> to copy the link.</p>
-      </DialogContent>
-      <DialogActions>
-        <Button colored type="button" onClick={closeDialog}>close</Button>
-      </DialogActions>
-    </Dialog>
-  );
-}
+import TrackListView from "./TrackListView";
+import Controls from "./Controls";
+import ShareDialog from "./ShareDialog";
 
 class App extends Component {
   loop: Tone.Sequence;
