+++ src/ShareDialog.js
@@ -1,0 +1,29 @@
+import React from 'react';
+import {
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogTitle
+} from "react-mdl";
+
+function ShareDialog({hash, closeDialog}) {
+  return (
+    <Dialog open>
+      <DialogTitle>Share</DialogTitle>
+      <DialogContent>
+        <p>Send this link to your friends so they can enjoy your piece:</p>
+        <p className="share-link" style={{textAlign: "center"}}>
+          <a className="mdl-button mdl-js-button mdl-button--colored"
+            href={"#" + hash} onClick={event => event.preventDefault()}>Link</a>
+        </p>
+        <p>Right-click, <em>Copy link address</em> to copy the link.</p>
+      </DialogContent>
+      <DialogActions>
+        <Button colored type="button" onClick={closeDialog}>close</Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
+
+export default ShareDialog;
